Completed angular material

+++ TaskManager/Controllers/TasksController.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
@@ -8,6 +9,7 @@ using TaskManager.Models;
 
 namespace TaskManager.Controllers
 {
+    [Authorize]
     [Route("api/Tasks")]
     [ApiController]
     public class TasksController : ControllerBase
@@ -21,80 +23,36 @@ namespace TaskManager.Controllers
 
         // GET: api/Tasks
         [HttpGet]
-        public IEnumerable<TaskModel> GetTasks([FromQuery] TaskParameters taskParameters)
+        public IActionResult GetTasks([FromQuery] TaskParameters taskParameters)
         {
             var taskDTOs = _service.GetTasks(taskParameters);
+            if (taskDTOs == null)
+            {
+                return NotFound();
+            }
             var mapper = new MapperConfiguration(cfg => cfg.CreateMap<TaskDTO, TaskModel>()).CreateMapper();
             var taskModels = mapper.Map<IEnumerable<TaskDTO>, List<TaskModel>>(taskDTOs);
-            return taskModels;
+            return Ok(taskModels);
         }
 
         // GET: api/Tasks/5
         [HttpGet("{id}")]
-        public TaskModel GetTask(int id)
+        public IActionResult GetTask(int id)
         {
             var taskDTO = _service.GetTask(id);
+            if (taskDTO == null)
+            {
+                return NotFound();
+            }
             var mapper = new MapperConfiguration(cfg => cfg.CreateMap<TaskDTO, TaskModel>()).CreateMapper();
             var taskModel = mapper.Map<TaskModel>(taskDTO);
-            return taskModel;
+            return Ok(taskModel);
         }
 
-        // GET: api/Tasks/finished
-        [HttpGet("finished")]
-        public IEnumerable<TaskModel> GetFinishedTasks()
-        {
-            var taskDTOs = _service.GetFinishedTasks();
-            var mapper = new MapperConfiguration(cfg => cfg.CreateMap<TaskDTO, TaskModel>()).CreateMapper();
-            var taskModels = mapper.Map<IEnumerable<TaskDTO>, List<TaskModel>>(taskDTOs);
-            return taskModels;
-        }
-
-
-        // GET: api/Tasks/unfinished
-        [HttpGet("unfinished")]
-        public IEnumerable<TaskModel> GetUnfinishedTasks()
-        {
-            var taskDTOs = _service.GetUnfinishedTasks();
-            var mapper = new MapperConfiguration(cfg => cfg.CreateMap<TaskDTO, TaskModel>()).CreateMapper();
-            var taskModels = mapper.Map<IEnumerable<TaskDTO>, List<TaskModel>>(taskDTOs);
-            return taskModels;
-        }
-
-
-        // GET: api/Tasks/waiting
-        [HttpGet("waiting")]
-        public IEnumerable<TaskModel> GetWaitingTasks()
-        {
-            var taskDTOs = _service.GetWaitingTasks();
-            var mapper = new MapperConfiguration(cfg => cfg.CreateMap<TaskDTO, TaskModel>()).CreateMapper();
-            var taskModels = mapper.Map<IEnumerable<TaskDTO>, List<TaskModel>>(taskDTOs);
-            return taskModels;
-        }
-
-        // GET: api/Tasks/unfinishedPassed
-        [HttpGet("unfinishedPassed")]
-        public IEnumerable<TaskModel> GetUnfinishedTasksWithPassedDeadline()
-        {
-            var taskDTOs = _service.GetUnfinishedTasksWithPassedDeadline();
-            var mapper = new MapperConfiguration(cfg => cfg.CreateMap<TaskDTO, TaskModel>()).CreateMapper();
-            var taskModels = mapper.Map<IEnumerable<TaskDTO>, List<TaskModel>>(taskDTOs);
-            return taskModels;
-        }
-
-        // GET: api/Tasks/unfinishedValid
-        [HttpGet("unfinishedValid")]
-        public IEnumerable<TaskModel> GetUnfinishedTasksWithValidDeadline()
-        {
-            var taskDTOs = _service.GetUnfinishedTasksWithValidDeadline();
-            var mapper = new MapperConfiguration(cfg => cfg.CreateMap<TaskDTO, TaskModel>()).CreateMapper();
-            var taskModels = mapper.Map<IEnumerable<TaskDTO>, List<TaskModel>>(taskDTOs);
-            return taskModels;
-        }
         // POST: api/Tasks
         [HttpPost]
-        public ActionResult<TaskModel> PostTask(TaskModel taskModel)
+        public IActionResult PostTask(TaskModel taskModel)
         {
-            var mapper = new MapperConfiguration(cfg => cfg.CreateMap<TaskModel, TaskDTO>()).CreateMapper();
             var newTask = new TaskDTO()
             {
                 Id = taskModel.Id,
@@ -106,16 +64,24 @@ namespace TaskManager.Controllers
             };
             _service.Create(newTask);
 
-            return CreatedAtAction("PostTask", new { id = newTask.Id }, newTask);
+            return Ok(new
+            {
+                Id = newTask.Id,
+                Name = newTask.Name
+            });
         }
 
         //PUT: api/Tasks/5
         [HttpPut("{id}")]
-        public ActionResult<TaskModel> PutTask(int id, TaskModel taskModel)
+        public IActionResult PutTask(int id, TaskModel taskModel)
         {
             var mapper = new MapperConfiguration(cfg => cfg.CreateMap<TaskModel, TaskDTO>()).CreateMapper();
             var taskDTO = mapper.Map<TaskDTO>(taskModel);
             var task = _service.GetTask(id);
+            if (task == null)
+            {
+                return NotFound();
+            }
             task.Name = taskDTO.Name;
             task.Status = taskDTO.Status;
             task.Created = taskDTO.Created;
@@ -128,13 +94,17 @@ namespace TaskManager.Controllers
 
         // DELETE: api/Tasks/5
         [HttpDelete("{id}")]
-        public ActionResult<TaskModel> DeleteTask(int id)
+        public IActionResult DeleteTask(int id)
         {
             var taskDTO = _service.GetTask(id);
+            if (taskDTO == null)
+            {
+                return NotFound();
+            }
             var mapper = new MapperConfiguration(cfg => cfg.CreateMap<TaskDTO, TaskModel>()).CreateMapper();
             var taskModel = mapper.Map<TaskModel>(taskDTO);
             _service.Delete(id);
-            return taskModel;
+            return Ok(taskModel);
         }
 
     }
